+++ packages/deploy/src/validator.ts
@@ -4,7 +4,7 @@ import { ProjectSpec } from './types';
 export interface Error {
   context: any;
   message: string;
-  path?: string[];
+  path: string; // human readable path to the error (ie, workflow-1/job-2)
   range?: [number, number, number];
 }
 
@@ -13,32 +13,42 @@ export function parseAndValidate(input: string): {
   doc: ProjectSpec;
 } {
   let errors: Error[] = [];
-  let keys: string[] = [];
   const doc = YAML.parseDocument(input);
 
-  function pushUniqueKey(context: YAML.Pair, key: string) {
-    if (keys.includes(key)) {
-      errors.push({
-        context,
-        message: `duplicate key: ${key}`,
-      });
+  function ensureUniqueId(key: string, arr: string[]) {
+    if (arr.includes(key)) {
+      throw `duplicate id: ${key}`;
     } else {
-      keys.push(key);
+      arr.push(key);
     }
   }
 
-  function validateJobs(workflow: YAMLMap) {
+  function validateJobs(
+    workflowName: string,
+    workflow: YAMLMap,
+    jobKeys: string[]
+  ) {
     const jobs = workflow.getIn(['jobs']);
 
     if (jobs) {
       if (isMap(jobs)) {
         for (const job of jobs.items) {
           if (isPair(job)) {
-            pushUniqueKey(job, (job as any).key.value);
+            const jobName = (job as any).key.value;
+            try {
+              ensureUniqueId(jobName, jobKeys);
+            } catch (err: any) {
+              errors.push({
+                path: `${workflowName}/${jobName}`,
+                context: job,
+                message: err,
+              });
+            }
           }
         }
       } else {
         errors.push({
+          path: 'workflow',
           context: jobs,
           message: 'jobs: must be a map',
         });
@@ -50,21 +60,39 @@ export function parseAndValidate(input: string): {
     if (typeof workflows === 'undefined') {
       // allow workflows to be unspecified, but ensure there is an empty
       // map to avoid errors downstream
-      doc.setIn(['workflows'], {})
-    }
-    else if (isMap(workflows)) {
+      doc.setIn(['workflows'], {});
+    } else if (isMap(workflows)) {
+      const workflowKeys: string[] = [];
       for (const workflow of workflows.items) {
         if (isPair(workflow)) {
-          pushUniqueKey(workflow, (workflow as any).key.value);
-
-          validateJobs((workflow as any).value);
+          const workflowName = (workflow as any).key.value;
+          const jobKeys: string[] = [];
+          try {
+            ensureUniqueId(workflowName, workflowKeys);
+          } catch (err: any) {
+            errors.push({
+              path: `${workflowName}`,
+              context: workflow,
+              message: err,
+            });
+          }
+          const workflowValue = (workflow as any).value;
+          if (isMap(workflowValue)) {
+            validateJobs(workflowName, workflowValue, jobKeys);
+          } else {
+            errors.push({
+              context: workflowValue,
+              message: `${workflowName}: must be a map`,
+              path: 'workflowName',
+            });
+          }
         }
       }
     } else {
       errors.push({
         context: workflows,
-        message: 'workflows: must be a map',
-        path: ['workflows'],
+        message: 'must be a map',
+        path: 'workflows',
       });
     }
   }
@@ -76,7 +104,7 @@ export function parseAndValidate(input: string): {
           errors.push({
             context: pair,
             message: 'project: must provide at least one workflow',
-            path: ['workflows'],
+            path: 'workflows',
           });
 
           return doc.createPair('workflows', {});
@@ -86,6 +114,7 @@ export function parseAndValidate(input: string): {
       if (pair.key && pair.key.value === 'jobs') {
         if (pair.value.value === null) {
           errors.push({
+            path: 'workflows',
             context: pair,
             message: 'jobs: must be a map',
             range: pair.value.range,
@@ -104,7 +133,11 @@ export function parseAndValidate(input: string): {
   });
 
   if (!doc.has('name')) {
-    errors.push({ context: doc, message: 'Project must have a name' });
+    errors.push({
+      context: doc,
+      message: 'Project must have a name',
+      path: 'project',
+    });
   }
 
   const workflows = doc.getIn(['workflows']);