+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @openfn/cli
 
+## 1.3.0
+
+### Minor Changes
+
+- 015055c: Add first pass of apollo command. Call an apollo service with `openfn apollo <service-name>`. For basic help run `openfn apollo --help`. For available services see the server index page. This first release is a super basic integration with log streaming through websockets and reasonably intelligent handling of `{ files }` result data.
+
+### Patch Changes
+
+- deploy: Improved error messages from local validation
+
+## 1.2.5
+
+### Patch Changes
+
+- Default the run timeout
+- Updated dependencies
+  - @openfn/runtime@1.1.3
+
+## 1.2.4
+
+### Patch Changes
+
+- deploy: Allow steps in different workflows to have the same name
+
 ## 1.2.3
 
 ### Patch Changes