+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @openfn/lightning-mock
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/engine-multi@1.1.8
+
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/runtime@1.1.3
+  - @openfn/engine-multi@1.1.7
+
 ## 2.0.6
 
 ### Patch Changes