+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,33 @@
 # ws-worker
 
+## 1.1.10
+
+### Patch Changes
+
+- bc45b3d: Restructure handling of env vars
+
+## 1.1.9
+
+### Patch Changes
+
+- Better error reporting for bad credentials
+- Updated dependencies
+  - @openfn/engine-multi@1.1.8
+
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/runtime@1.1.3
+  - @openfn/engine-multi@1.1.7
+
+## 1.1.7
+
+### Patch Changes
+
+- bdff4b2: Fix an issue where workers may not be returned to the pool if run:complete throws
+
 ## 1.1.6
 
 ### Patch Changes